bot_compose/bot/requirements.txt
aiogram==2.14.3
pydantic==1.10.2
python-dotenv==0.21.0
loguru==0.6.0
aioredis<2.0